+++ src/main/java/com/garden/coffee/coffeegarden/product/ProductDto.java
@@ -7,8 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.Column;
-import javax.persistence.Id;
+import jakarta.persistence.Column;
+import jakarta.persistence.Id;
 import java.security.NoSuchAlgorithmException;
 
 @Getter
@@ -17,10 +17,10 @@ import java.security.NoSuchAlgorithmException;
 public class ProductDto implements DataTransferObject<Product> {
     public String productId;
     public String productName;
-    public Long productPrice;
-    public Long salePercent;
+    public int productPrice;
+    public int salePercent;
 
-    public ProductDto(String productId,String productName, Long productPrice, Long salePercent){
+    public ProductDto(String productId,String productName, int productPrice, int salePercent){
         this.productId = productId;
         this.productName = productName;
         this.productPrice = productPrice;
